requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.20
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-import-export==2.5.0
django-js-reverse==0.9.1
django-model-utils==4.1.1
django-rest-swagger==2.2.0
django-webpack-loader==0.7.0
djangorestframework
et-xmlfile==1.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn==19.9.0
django-taggit
dynamic-rest==2.0.0
djangorestframework-queryfields==1.0.0